+++ src/utils/constants/constant.js
@@ -1,6 +1,8 @@
 export const PATH_HOME = "/";
 export const PATH_DASHBOARD = "/dashboard";
+export const PATH_PAYMENTS = "/payments";
 export const PATH_TRANSACTIONS = "/transactions";
+
 
 export const LOCALSTORAGE_ACCESS_TOKEN = process.env.REACT_APP_LOCALSTORAGE_ACCESS_TOKEN;
 export const LOCALSTORAGE_ID_TOKEN = process.env.REACT_APP_LOCALSTORAGE_ID_TOKEN;
@@ -19,4 +21,6 @@ export const LOADING = "loading";
 
 export const LOGIN_SHARED_KEY = "loginSharedKey";
 
+export const ENTITY_LOGIN = "login";
+export const ENTITY_PAYMENT = "payment";
 export const ENTITY_TRANSACTIONS = "transactions";